Move testing platform targets into separate directories
