+++ packages/constraint-engine/src/customRules/isObject.ts
@@ -1,4 +1,4 @@
-import { ObjectTarget } from '../customTargets/testingPlatform/targets';
+import { ObjectTarget } from '../customTargets/type-script/objectTarget';
 import { Rule } from '../types/rule';
 
 export const isObject = ((target: unknown): target is ObjectTarget => {
