+++ packages/constraint-engine/src/customTargets/testingPlatform/packageC/packageConfigurationTarget.ts
@@ -1,0 +1,8 @@
+export enum PackageConfigurationTypeId {
+  TestFramework = 'TestFramework',
+  NonTestFramework = 'NonTestFramework',
+}
+
+export type TestingPlatformConfigurationTarget = {
+  typeId: PackageConfigurationTypeId;
+};
