+++ packages/constraint-engine/src/customTargets/testingPlatform/packageDirectory/buildTestingPlatformPackageDirectoryReferenceSet.test.ts
@@ -3,8 +3,8 @@ import {
   report,
   signaler,
 } from 'rat-test/type-script/transgressing';
+import { TestingPlatformTargetTypeId } from '../targetTypeIds';
 import { buildTestingPlatformPackageDirectoryReferenceSet } from './buildTestingPlatformPackageDirectoryReferenceSet';
-import { TestingPlatformTargetTypeId } from '../targets';
 
 // eslint-disable-next-line @typescript-eslint/no-floating-promises
 orchestrate()
