+++ packages/constraint-engine/src/customTargets/testingPlatform/packageDirectory/packageDirectoryTarget.ts
@@ -1,0 +1,12 @@
+import { TypedTarget } from '../../../types/typedTarget';
+import { TestingPlatformTargetTypeId } from '../targetTypeIds';
+
+export type TestingPlatformPackageDirectoryTarget = {
+  directoryPath: string;
+  index: number;
+};
+
+export type TestingPlatformPackageDirectoryTypedTarget = TypedTarget<
+  TestingPlatformTargetTypeId.PackageDirectory,
+  TestingPlatformPackageDirectoryTarget
+>;
