+++ packages/constraint-engine/src/customTargets/testingPlatform/packageDirectorySet/packageDirectorySetTarget.ts
@@ -1,0 +1,9 @@
+import { TypedTarget } from '../../../types/typedTarget';
+import { TestingPlatformTargetTypeId } from '../targetTypeIds';
+
+export type TestingPlatformPackageDirectorySetTarget = string[];
+
+export type TestingPlatformPackageDirectorySetTypedTarget = TypedTarget<
+  TestingPlatformTargetTypeId.PackageDirectorySet,
+  TestingPlatformPackageDirectorySetTarget
+>;
