+++ packages/constraint-engine/src/customTargets/testingPlatform/targetTypeIds.ts
@@ -1,0 +1,7 @@
+export enum TestingPlatformTargetTypeId {
+  PackageDirectorySet = 'PackageDirectorySet',
+  PackageDirectory = 'PackageDirectory',
+  PackageA = 'PackageA',
+  PackageB = 'PackageB',
+  PackageC = 'PackageC',
+}
