+++ packages/constraint-engine/src/customTargets/type-script/objectTarget.ts
@@ -1,0 +1,1 @@
+export type ObjectTarget = Record<string, unknown>;
